Add authorization

+++ routes/campgrounds.js
@@ -1,20 +1,8 @@
 const express = require("express");
 const router = express.Router();
 const catchAsync = require("../utils/catchAsync");
-const ExpressError = require("../utils/ExpressError");
-const { campgroundSchema } = require("../schemas.js");
 const Campground = require("../models/campground");
-const { isLoggedIn } = require("../middleware");
-
-const validateCampground = (req, res, next) => {
-  const { error } = campgroundSchema.validate(req.body);
-  if (error) {
-    const msg = error.details.map((el) => el.message).join(",");
-    throw new ExpressError(msg, 400);
-  } else {
-    next();
-  }
-};
+const { isLoggedIn, isAuthor, validateCampground } = require("../middleware");
 
 router.get(
   "/",
@@ -33,9 +21,6 @@ router.post(
   isLoggedIn,
   validateCampground,
   catchAsync(async (req, res, next) => {
-    // if (!req.body.campground)
-    //   throw new ExpressError("Invalid Campground Data", 400);
-
     const campground = new Campground(req.body.campground);
     campground.author = req.user._id;
     await campground.save();
@@ -48,8 +33,8 @@ router.get(
   "/:id",
   catchAsync(async (req, res) => {
     const campground = await Campground.findById(req.params.id)
-      .populate("reviews")
-      .populate("author");
+      .populate({ path: "reviews", populate: { path: "author" } })
+      .populate({ path: "author" });
     console.log(campground);
     if (!campground) {
       req.flash("error", "Cannot find that campground...");
@@ -62,8 +47,10 @@ router.get(
 router.get(
   "/:id/:edit",
   isLoggedIn,
+  isAuthor,
   catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
     if (!campground) {
       req.flash("error", "Cannot find that campground...");
       return res.redirect("/campgrounds");
@@ -75,12 +62,14 @@ router.get(
 router.put(
   "/:id",
   isLoggedIn,
+  isAuthor,
   validateCampground,
   catchAsync(async (req, res) => {
     const { id } = req.params;
     const campground = await Campground.findByIdAndUpdate(id, {
       ...req.body.campground,
     });
+    req.flash("success", "Successfully updated!");
     res.redirect(`/campgrounds/${campground._id}`);
   })
 );
@@ -90,6 +79,14 @@ router.delete(
   isLoggedIn,
   catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash("error", "You do not have permission to do that");
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndUpdate(id, {
+      ...req.body.campground,
+    });
     await Campground.findByIdAndDelete(id);
     req.flash("success", "Successfully deleted a campground!");
     res.redirect("/campgrounds");
